requirements.txt
django-allauth

Django

gunicorn

django-bootstrap-v5

django-bootstrap-datepicker-plus

django-heroku

python-dotenv

django-storages

boto3

Pillow

django_cron

twilio

django-star-ratings